terraform/functions/join_meeting/requirements.txt
google-cloud-pubsub==2.18.4
google-cloud-firestore==2.13.1
google-cloud-logging==3.8.0
